Refactor doubles into BigDecimal numbers
Doubles are notoriously imprecise for computations. This commit now
utilizes BigDoubles with high precision for derivative computations.

+++ src/main/java/node/VarNode.java
@@ -2,6 +2,7 @@ package node;
 
 import exceptions.UndefinedVariable;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 /**
@@ -17,8 +18,8 @@ public class VarNode extends OperandNode
 	}
 
 	@Override
-	public double eval(Map<String, Double> env) throws UndefinedVariable {
-		Double thisVal = env.get(variable + "");
+	public BigDecimal eval(Map<String, BigDecimal> env) throws UndefinedVariable {
+		BigDecimal thisVal = env.get(variable + "");
 		if (thisVal == null)
 			throw new UndefinedVariable("The variable: " + variable + " is not defined in the env.");
 		return thisVal;
